Move exceptions to a separate module
Move exceptions out to a separate module so they can be easily imported
without bringing in all of the Base module.
